+++ R/Api.R
@@ -2,75 +2,102 @@
 # Author: Sam Hausmann
 
 library(httr)
-library(rjson)
+library(jsonlite)
 
 #' api wrapper function that checks for a user_key and determines the Rosette API endpoint to be utilized
 #' @param user_key - Rosette API authentication key
 #' @param endpoint - Rosette API endpoint to be utilized
 #' @param parameters - parameters list to be passed to specified Rosette API endpoint
-#' @param customHeaders - custom headers for Rosette Api
+#' @param custom_headers - custom headers for Rosette Api
+#' @param url_parameters - query parameters
 #' @param url - url for Rosette Api
-#' @return Returns json of the specified Rosette API endpoint response
+#' @return Returns a list(content, header)
 #' @examples
 #' \dontrun{
-#' key <- "content"
-#' value <- "Bill Murray will appear in new Ghostbusters film."
 #' parameters <- list()
-#' parameters[[ key ]] <- value
-#' parameters <- toJSON(parameters)
-#' api(01234567890, "entities", parameters)
+#' parameters[[ "content" ]] <- "Bill Murray will appear in new Ghostbusters film."
+#' response = api(01234567890, "entities", parameters) will return response$content and response$header
 #' }
 #' @export
-api <- function(user_key, endpoint, parameters=FALSE, customHeaders=NULL, url="https://api.rosette.com/rest/v1/") {
-  if(is.null(user_key)) {
+api <- function(user_key, endpoint, parameters=FALSE, custom_headers=NULL, url_parameters=NULL, url="https://api.rosette.com/rest/v1/") {
+  if (is.null(user_key)) {
     stop("API key param empty")
   } else {
-
-    if(endpoint == "info") {
-      return(to_json(get_endpoint(user_key, "info", customHeaders, url)))
-    } else if (endpoint == "ping") {
-      return(to_json(get_endpoint(user_key, "ping", customHeaders, url)))
-    } else if (endpoint == "language") {
-      return(to_json(error_check(check_for_multipart(user_key, check_content_parameters(parameters), "language", customHeaders, url))))
-    } else if (endpoint == "categories") {
-      return(to_json(error_check(check_for_multipart(user_key, check_content_parameters(parameters), "categories", customHeaders, url))))
-    } else if (endpoint == "entities") {
-      return(to_json(error_check(check_for_multipart(user_key, check_content_parameters(parameters), "entities", customHeaders, url))))
-    } else if (endpoint == "name-translation") {
-      return(to_json(error_check(check_for_multipart(user_key, check_names(parameters, "name-translation"), "name-translation", customHeaders, url))))
-    } else if (endpoint == "name-similarity") {
-      return(to_json(error_check(check_for_multipart(user_key, check_names(parameters, "name-similarity"), "name-similarity", customHeaders, url))))
-    } else if (endpoint  == "relationships") {
-      return(to_json(error_check(check_for_multipart(user_key, check_content_parameters(parameters), "relationships", customHeaders, url))))
-    } else if (endpoint == "tokens") {
-      return(to_json(error_check(check_for_multipart(user_key, check_content_parameters(parameters), "tokens", customHeaders, url))))
-    } else if (endpoint == "morphology") {
-      return(to_json(error_check(check_for_multipart(user_key, check_content_parameters(parameters), check_morphology(parameters, "morphology"), customHeaders, url))))
-    } else if (endpoint == "sentiment") {
-      return(to_json(error_check(check_for_multipart(user_key, check_content_parameters(parameters), "sentiment", customHeaders, url))))
-    } else if (endpoint == "sentences") {
-      return(to_json(error_check(check_for_multipart(user_key, check_content_parameters(parameters), "sentences", customHeaders, url))))
-    } else if (endpoint == "entities/linked"){
-      .Deprecated(api(user_key, "entities"), NULL, "/entities/linked is deprecated. The endpoint has been merged with /entities")
-      return(to_json(error_check(check_for_multipart(user_key, check_content_parameters(parameters), "entities/linked", customHeaders, url))))
-    } else if (endpoint == "text-embedding"){
-      return(to_json(error_check(check_for_multipart(user_key, check_content_parameters(parameters), "text-embedding", customHeaders, url))))
-    } else {
-      stop("Specify a valid Rosette API endpoint")
-
+    if (!is.null(custom_headers)) {
+      if (grepl("^X-RosetteAPI-", names(custom_headers)) == FALSE){
+        stop("Custom headers must start with \"X-RosetteAPI-\"")
+      }
+    }
+    if (!endsWith(url, "/")) {
+      url <- paste(url, "/", sep = "")
     }
 
-  }
 
+
+    response <- switch(endpoint,
+      "info"= error_check(
+        get_endpoint(user_key, "info", url, custom_headers, url_parameters)
+      ),
+      "ping"= error_check(
+        get_endpoint(user_key, "ping", url, custom_headers, url_parameters)
+      ),
+      "language"= error_check(
+        post_endpoint(user_key, check_content_parameters(parameters), "language", url, custom_headers, url_parameters)
+      ),
+      "categories"= error_check(
+        post_endpoint(user_key, check_content_parameters(parameters), "categories", url, custom_headers, url_parameters)
+      ),
+      "entities"= error_check(
+        post_endpoint(user_key, check_content_parameters(parameters), "entities", url, custom_headers, url_parameters)
+      ),
+      "name-translation"= error_check(
+        post_endpoint(user_key, check_names(parameters, "name-translation"), "name-translation", url, custom_headers, url_parameters)
+      ),
+      "name-similarity"= error_check(
+        post_endpoint(user_key, check_names(parameters, "name-similarity"), "name-similarity", url, custom_headers, url_parameters)
+      ),
+      "relationships"= error_check(
+        post_endpoint(user_key, check_content_parameters(parameters), "relationships", url, custom_headers, url_parameters)
+      ),
+      "tokens"= error_check(
+        post_endpoint(user_key, check_content_parameters(parameters), "tokens", url, custom_headers, url_parameters)
+      ),
+      "morphology"= error_check(
+        post_endpoint(user_key, check_content_parameters(parameters), check_morphology(parameters, "morphology"), url, custom_headers, url_parameters)
+      ),
+      "sentiment"= error_check(
+        post_endpoint(user_key, check_content_parameters(parameters), "sentiment", url, custom_headers, url_parameters)
+      ),
+      "sentences"= error_check(
+        post_endpoint(user_key, check_content_parameters(parameters), "sentences", url, custom_headers, url_parameters)
+      ),
+      "text-embedding"= error_check(
+        post_endpoint(user_key, check_content_parameters(parameters), "text-embedding", url, custom_headers, url_parameters)
+      ),
+      "syntax/dependencies"= error_check(
+        post_endpoint(user_key, check_content_parameters(parameters), "syntax/dependencies", url, custom_headers, url_parameters)
+      ),
+      stop("Specify a valid Rosette API endpoint")
+    )
+
+    return(list("content" = httr::content(response), "header" = httr::headers(response)))
+  }
+}
+
+#' Provides the binding Version
+#' @return current binding version
+get_binding_version <- function() {
+  bindingVersion <- "1.5.0"
+  return(bindingVersion)
 }
 
 #' preemptive check for content/contentUri request parameter errors
 #' @param parameters - parameters list to be passed to specified Rosette API endpoint
 #' @return Returns list of verified parameters to be sent to Rosette API
 check_content_parameters <- function(parameters) {
-  if("content" %in% names(fromJSON(parameters))  && "contentUri" %in% names(fromJSON(parameters))){
+  if ("content" %in% names(parameters)  && "contentUri" %in% names(parameters)){
     stop("Cannot specify both content and contentUri parameters")
-  } else if(!("content" %in% names(fromJSON(parameters)))  && !("contentUri" %in% names(fromJSON(parameters)))) {
+  } else if (!("content" %in% names(parameters))  && !("contentUri" %in% names(parameters))) {
     stop("Must specify either content or contentUri parameters")
   } else {
     return(parameters)
@@ -82,8 +109,8 @@ check_content_parameters <- function(parameters) {
 #' @param endpoint - Rosette API endpoint to be utilized
 #' @return Returns the specified morphology function
 check_morphology <- function(parameters, endpoint) {
-  if("morphology" %in% names(fromJSON(parameters))) {
-    endpoint <- paste("morphology/", fromJSON(parameters)["morphology"], sep="")
+  if ("morphology" %in% names(parameters)) {
+    endpoint <- paste("morphology/", parameters["morphology"], sep="")
     return(endpoint)
   } else {
     return(endpoint)
@@ -95,14 +122,15 @@ check_morphology <- function(parameters, endpoint) {
 #' @param endpoint - Rosette API endpoint to be utilized
 #' @return Returns list of verified parameters to be sent to Rosette API
 check_names <- function(parameters, endpoint) {
-  if(endpoint == "name-translation") {
-    if(!("name" %in% names(fromJSON(parameters))) || !("targetLanguage" %in% names(fromJSON(parameters)))) {
+  params <- parameters
+  if (endpoint == "name-translation") {
+    if (!("name" %in% names(params)) || !("targetLanguage" %in% names(params))) {
       stop("Must supply name and targetLanguage parameters for name-translation endpoint")
     } else {
       return(parameters)
     }
-  } else if(endpoint == "name-similarity") {
-    if(!("name1" %in% names(fromJSON(parameters))) || !("name2" %in% names(fromJSON(parameters)))) {
+  } else if (endpoint == "name-similarity") {
+    if (!("name1" %in% names(params)) || !("name2" %in% names(params))) {
       stop("Must supply both name1 and name2 parameters for name-similarity endpoint")
     } else {
       return(parameters)
@@ -110,30 +138,13 @@ check_names <- function(parameters, endpoint) {
   }
 }
 
-#' check if the request parameters contain multipart
-#' @param user_key - Rosette API authentication key
-#' @param endpoint - Rosette API endpoint to be utilized
-#' @param parameters - parameters list to be passed to specified Rosette API endpoint
-#' @param customHeaders - custom headers for Rosette Api
-#' @param url - url for Rosette Api
-#' @return Returns json of the specified Rosette API endpoint response from either a multipart or non-multipart request
-check_for_multipart <- function(user_key, parameters, endpoint, customHeaders, url) {
-  if("documentFile" %in% names(fromJSON(parameters))) {
-    create_multipart(parameters)
-    return(mutipart_call(user_key, create_multipart(parameters), endpoint, customHeaders, url))
-  } else {
-    return(post_endpoint(user_key, parameters, endpoint, customHeaders, url))
-  }
-}
-
 #' create a multipart
 #' @param parameters - parameters list to be passed to specified Rosette API endpoint
 #' @return Returns a multipart
 create_multipart <- function(parameters) {
-  content <- fromJSON(parameters)['content']
-  parameters <- fromJSON(parameters)
+  content <- parameters['content']
   parameters['content'] <- NULL
-  parameters <- serialize_parameters(toJSON(parameters))
+  parameters <- serialize_parameters(parameters)
 
   boundary <- "--89dszpjalrbmlsor"
   crlf <- "\r\n"
@@ -143,7 +154,7 @@ create_multipart <- function(parameters) {
   multi <- paste(multi, 'Content-Disposition: mixed; name="request"', sep='')
   multi <- paste(multi, crlf, sep='')
   multi <- paste(multi, crlf, sep='')
-  if(length(fromJSON(parameters)) != 0){
+  if (length(jsonlite::fromJSON(parameters)) != 0){
     multi <- paste(multi, parameters, sep='')
   } else {
     multi <- paste(multi, "{}", sep='')
@@ -164,46 +175,28 @@ create_multipart <- function(parameters) {
   return(multi)
 }
 
-#' make a multipart POST request
-#' @param user_key - Rosette API authentication key
-#' @param endpoint - Rosette API endpoint to be utilized
-#' @param parameters - parameters list to be passed to specified Rosette API endpoint
-#' @param customHeaders - custom headers for Rosette Api
-#' @param url - url for Rosette Api
-#' @return Returns the response from the Rosette API
-mutipart_call <- function(user_key, parameters, endpoint, customHeaders,url) {
-  BINDING_VERSION <- "1.3.0"
-  if(is.null(customHeaders)) {
-    response <- POST(paste(url, endpoint, sep=""), encode = "multipart", add_headers("X-RosetteAPI-Key" = user_key, "Content-Type" = "multipart/mixed", "X-RosetteAPI-Binding" = "R", "X-RosetteAPI-Binding-Version" = BINDING_VERSION, "user-agent" = "ruseragent"), body = parameters)
-    } else {
-        defaultHeaders <- c("X-RosetteAPI-Key" = user_key, "Content-Type" = "multipart/mixed", "X-RosetteAPI-Binding" = "R", "X-RosetteAPI-Binding-Version" = BINDING_VERSION, "user-agent" = "ruseragent")
-        .headers <- c(defaultHeaders, customHeaders)
-        response <- POST(paste(url, endpoint, sep=""), encode = "multipart", add_headers(.headers), body = parameters)
-    }
-  return(response)
-}
-
-#' convert Rosette API response content and headers to json
-#' @param response - response from Rosette API
-#' @return Returns a vector containing json of Rosette API response and response headers
-to_json <- function(response) {
-  resultVector = vector()
-  json <- toJSON(content(response))
-  headers <- toJSON(headers(response))
-  resultVector[1] <- json
-  resultVector[2] <- headers
-  return(resultVector)
-}
-
 #' check if Rosette API response includes and error message
 #' @param response - response from Rosette API
 #' @return Returns an error if one exists or the Rosette API response
 error_check <- function(response) {
-  if("message" %in% names(content(response))) {
-    stop(content(response)["message"])
+  if ("message" %in% names(response)) {
+    stop(response["message"])
   } else {
     return(response)
   }
+}
+
+#' Helper to combine custom headers with the default ones
+#' @param user_key - Rosette API authentication key
+#' @param content_type - Header Content-Type
+#' @param custom_headers - custom headers for the Rosette API
+#' @return Returns the combined h
+get_headers <- function(user_key, content_type="application/json", custom_headers=NULL) {
+  headers <- c("X-RosetteAPI-Key" = user_key, "Content-Type" = content_type, "X-RosetteAPI-Binding" = "R", "X-RosetteAPI-Binding-Version" = get_binding_version(), "user-agent" = "ruseragent")
+  if (!is.null(custom_headers)) {
+    return(c(headers, custom_headers))
+  }
+  return(headers)
 }
 
 #' serialize Rosette API parameters
@@ -211,12 +204,12 @@ error_check <- function(response) {
 #' @return Returns the serialized parameters for the Rosette API
 serialize_parameters <- function(parameters) {
   serialized_params <- list()
-  for(param in names(fromJSON(parameters))) {
-    if(param == "genre" || param == "profileId" || param == "language" || param == "content" || param == "options" || param == "contentUri" || param == "contentType") {
-    serialized_params[[param]] <- fromJSON(parameters)[[param]]
+  for (param in names(parameters)) {
+    if (param == "genre" || param == "profileId" || param == "language" || param == "content" || param == "options" || param == "contentUri" || param == "content_type") {
+    serialized_params[[param]] <- parameters[[param]]
     }
   }
-  return(toJSON(serialized_params))
+  return(jsonlite::toJSON(serialized_params, auto_unbox = TRUE))
 }
 
 #' serialize Rosette API parameters
@@ -224,63 +217,83 @@ serialize_parameters <- function(parameters) {
 #' @return Returns the serialized parameters for the Rosette API
 serialize_name_parameters <- function(parameters) {
   serialized_params <- list()
-  for(param in names(fromJSON(parameters))) {
-    if(param == "name" || param == "targetLanguage" || param == "targetScript" || param == "entityType" || param == "sourceScript" || param == "sourceLanguageOfOrigin" || param == "sourceLanguageOfUse" || param == "targetScheme" || param == "name1" || param == "name2") {
-    serialized_params[[param]] <- fromJSON(parameters)[[param]]
+  for (param in names(parameters)) {
+    if (param == "name" || param == "targetLanguage" || param == "targetScript" || param == "entityType" || param == "sourceScript" || param == "sourceLanguageOfOrigin" || param == "sourceLanguageOfUse" || param == "targetScheme" || param == "name1" || param == "name2") {
+    serialized_params[[param]] <- parameters[[param]]
     }
   }
-  return(toJSON(serialized_params))
+  return(jsonlite::toJSON(serialized_params, auto_unbox = TRUE))
 }
 
-#' POST request to specified Rosette API endpoint
+#' Helper to check for file submission
+#' @param parameters - JSON parameters
+#' @return true if multipart
+check_for_multipart <- function(parameters) {
+  return("documentFile" %in% names(parameters))
+}
+
+#' httr::POST request to specified Rosette API endpoint
 #' @param user_key - Rosette API authentication key
 #' @param endpoint - Rosette API endpoint to be utilized
 #' @param parameters - parameters list to be passed to specified Rosette API endpoint
-#' @param customHeaders - custom headers for Rosette Api
 #' @param url - url for Rosette Api
-#' @return Returns the response from the Rosette API
-post_endpoint <- function(user_key, parameters, endpoint, customHeaders, url) {
-  BINDING_VERSION <- "1.3.0"
-  if(is.null(customHeaders)) {
-      if(endpoint == "name-translation" || endpoint == "name-similarity"){
-        response <- POST(paste(url, endpoint, sep=""), add_headers("X-RosetteAPI-Key" = user_key, "Content-Type" = "application/json", "X-RosetteAPI-Binding" = "R", "X-RosetteAPI-Binding-Version" = BINDING_VERSION), body = serialize_name_parameters(parameters))
-      } else {
-        response <- POST(paste(url, endpoint, sep=""), add_headers("X-RosetteAPI-Key" = user_key, "Content-Type" = "application/json", "X-RosetteAPI-Binding" = "R", "X-RosetteAPI-Binding-Version" = BINDING_VERSION), body = serialize_parameters(parameters))
-      }
+#' @param custom_headers - custom headers for Rosette Api
+#' @param url_parameters - query parameters
+#' @return Returns the Rosette API response
+post_endpoint <- function(user_key, parameters, endpoint, url, custom_headers=NULL, url_parameters=NULL) {
+  if (check_for_multipart(parameters)) {
+    request_body <- create_multipart(parameters)
+    encoding <- "multipart"
+    content_type = "multipart/mixed"
+  } else {
+    encoding <- "json"
+    content_type <- "application/json"
+    if (endpoint == "name-translation" || endpoint == "name-similarity") {
+      request_body <- serialize_name_parameters(parameters)
     } else {
-      if(grepl("^X-RosetteAPI-", names(customHeaders)) == FALSE){
-        stop("Custom headers must start with \"X-\"")
-        } else {
-          defaultHeaders <- c("X-RosetteAPI-Key" = user_key, "Content-Type" = "application/json", "X-RosetteAPI-Binding" = "R", "X-RosetteAPI-Binding-Version" = BINDING_VERSION, "user-agent" = "ruseragent")
-          .headers <- c(defaultHeaders, customHeaders)
-          if(endpoint == "name-translation" || endpoint == "name-similarity"){
-            response <- POST(paste(url, endpoint, sep=""), add_headers(.headers), body = serialize_name_parameters(parameters))
-          } else {
-            response <- POST(paste(url, endpoint, sep=""), add_headers(.headers), body = serialize_parameters(parameters))
-          }
-        }
+      request_body <- serialize_parameters(parameters)
     }
+  }
+
+  if (is.null(url_parameters)) {
+    response <- httr::POST(
+      paste(url, endpoint, sep=""), 
+      encode = encoding,
+      httr::add_headers(get_headers(user_key, content_type = content_type, custom_headers = custom_headers)), 
+      body = request_body
+    )
+  } else {
+    response <- httr::POST(
+      paste(url, endpoint, sep=""), 
+      encode = encoding,
+      httr::add_headers(get_headers(user_key, content_type = content_type, custom_headers = custom_headers)), 
+      body = request_body,
+      query = url_parameters
+    )
+  }
   return(response)
 }
 
-#' GET request to specified Rosette API endpoint
+#' httr::GET request to specified Rosette API endpoint
 #' @param user_key - Rosette API authentication key
 #' @param endpoint - Rosette API endpoint to be utilized
-#' @param customHeaders - custom headers for Rosette Api
 #' @param url - url for Rosette Api
-#' @return Returns the response from the Rosette API
-get_endpoint <- function(user_key, endpoint, customHeaders, url) {
-  BINDING_VERSION <- "1.3.0"
-  if(is.null(customHeaders)) {
-    response <- GET(paste(url, endpoint, sep=""), add_headers("X-RosetteAPI-Key" = user_key, "X-RosetteAPI-Binding" = "R", "X-RosetteAPI-Binding-Version" = BINDING_VERSION, "user-agent" = "ruseragent"))
-    } else {
-      if(grepl("^X-RosetteAPI-", names(customHeaders)) == FALSE){
-        stop("Custom headers must start with \"X-\"")
-        } else {
-          defaultHeaders <- c("X-RosetteAPI-Key" = user_key, "Content-Type" = "application/json", "X-RosetteAPI-Binding" = "R", "X-RosetteAPI-Binding-Version" = BINDING_VERSION, "user-agent" = "ruseragent")
-          .headers <- c(defaultHeaders, customHeaders)
-          response <- GET(paste(url, endpoint, sep=""), add_headers(.headers))
-        }
-    }
+#' @param custom_headers - custom headers for Rosette Api
+#' @param url_parameters - url query parameters
+#' @return Returns the Rosette API response
+get_endpoint <- function(user_key, endpoint, url, custom_headers=NULL, url_parameters=NULL) {
+  if (is.null(url_parameters)) {
+    response <- httr::GET(
+      paste(url, endpoint, sep=""), 
+      httr::add_headers(get_headers(user_key, custom_headers))
+    )
+  } else {
+    response <- httr::GET(
+      paste(url, endpoint, sep=""), 
+      httr::add_headers(get_headers(user_key, custom_headers)),
+      query = url_parameters
+    )
+  }
+
   return(response)
 }