+++ examples/ping.R
@@ -1,15 +1,17 @@
 source("../R/Api.R")
-library(rjson)
+library(jsonlite)
 library("optparse")
 
 option_list = list( make_option(c("-k", "--key"), action="store", default=NA, type='character',
-              help="Rosette API key"))
+              help="Rosette API key"), make_option(c("-u", "--url"), action="store", default=NA, type='character',
+              help="Rosette API url"))
 opt_parser = OptionParser(option_list=option_list)
 opt = parse_args(opt_parser)
 
-if(is.null(opt$url)){
-    result <- api(opt$key, "ping")
+if (is.na(opt$url)){
+   result <- api(opt$key, "ping")
 } else {
-    result <- api(opt$key, "ping", NULL, NULL, opt$url)
+   result <- api(opt$key, "ping", NULL, NULL, opt$url)
 }
-print(result)
+print(jsonlite::toJSON(result$header, pretty = TRUE))
+print(jsonlite::toJSON(result$content, pretty = TRUE))