+++ examples/text_embedding.R
@@ -1,5 +1,5 @@
 source("../R/Api.R")
-library(rjson)
+library(jsonlite)
 library("optparse")
 
 option_list = list( make_option(c("-k", "--key"), action="store", default=NA, type='character',
@@ -10,16 +10,13 @@ opt = parse_args(opt_parser)
 
 embeddings_data <- "Cambridge, Massachusetts"
 
-key <- "content"
-value <- entities_linked_text_data
-
 parameters <- list()
-parameters[[ key ]] <- value
-parameters <- toJSON(parameters)
+parameters[[ "content" ]] <- embeddings_data
 
-if(is.na(opt$url)){
+if (is.na(opt$url)){
    result <- api(opt$key, "text-embedding", parameters)
 } else {
    result <- api(opt$key, "text-embedding", parameters, NULL, opt$url)
 }
-print(result)
+print(jsonlite::toJSON(result$header, pretty = TRUE))
+print(jsonlite::toJSON(result$content, pretty = TRUE))