+++ man/NEWS.md
@@ -11,4 +11,18 @@
 - Parameters serialized correctly
 
 # rosetteApi 1.3.0
-- New text-embedding example
+- New text_embedding example
+
+# rosetteApi 1.4.0
+- New syntax/dependencies endpoint
+- New syntax_dependencies example
+
+# rosetteApi 1.4.1
+- Removed deprecated entities linked
+
+# rosetteApi 1.4.2
+- Refactor of internal code
+- Modified api response to be a list(content, header)
+- Changed from rjson to jsonlite
+- Cleaned up examples
+- parameter argument is now a list and does not need to be converted to JSON